+++ addOns/client/CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Add a button in the Client History tab to clear the history from both the GUI and session.
 
 ### Changed
+- Allow callback implementors to handle browsers closing.
 - Depend on Database add-on.
 
 ## [0.21.0] - 2026-03-31